requirements.txt
pandas==2.0.1
python-dateutil==2.8.2
numpy==1.26.4
pyyaml==6.0.2
pytest==8.3.5